refactor: update logging methods in audit trail classes for consistency
