+++ Hermodr.sln
@@ -704,11 +704,9 @@ Global
 		{397F8FF6-E1BE-44AB-863A-7EDD8BDF6B19} = {537B0F0D-AC12-49A9-B54B-21456E9139CB}
 		{20DE5261-A9DB-4376-B1FE-B17CFFA7419D} = {EDED427E-1408-4971-9FA9-EBFCACCEAC22}
 		{E7651255-1279-41EC-88A9-6ADA92B4C544} = {EDED427E-1408-4971-9FA9-EBFCACCEAC22}
-		{1B65E7C5-1857-48DB-9F8D-07F67C39FA0B} = {B34ECEB3-D2F9-459B-A3DE-99E04123E593}
 		{F6B5796B-C891-4D23-9493-47C718C28B76} = {EDED427E-1408-4971-9FA9-EBFCACCEAC22}
 		{1BA69448-5794-4A53-B010-B0DB3D0C2CC1} = {B34ECEB3-D2F9-459B-A3DE-99E04123E593}
 		{1C9AA5DC-E964-41EF-BAF7-7579FC94C3D6} = {B34ECEB3-D2F9-459B-A3DE-99E04123E593}
-		{F21CD9C4-9B97-4E5D-8E88-1274931C8742} = {B34ECEB3-D2F9-459B-A3DE-99E04123E593}
 		{7A74A140-51AE-4B78-939D-7906D139D245} = {B34ECEB3-D2F9-459B-A3DE-99E04123E593}
 		{3560930F-0B81-4580-A053-756E79A2CA66} = {EDED427E-1408-4971-9FA9-EBFCACCEAC22}
 		{A2D18273-EA79-4AB8-856C-56BCBB3C6865} = {EDED427E-1408-4971-9FA9-EBFCACCEAC22}
@@ -716,6 +714,8 @@ Global
 		{F76555CF-03AE-449D-A460-E1A63A29E11C} = {EDED427E-1408-4971-9FA9-EBFCACCEAC22}
 		{6ABD28A3-CE97-2F60-AC77-E60C09A9F147} = {5D20AA90-6969-D8BD-9DCD-8634F4692FDA}
 		{337123F6-A3FD-44AA-856E-B38513EEDA67} = {6ABD28A3-CE97-2F60-AC77-E60C09A9F147}
+		{F21CD9C4-9B97-4E5D-8E88-1274931C8742} = {537B0F0D-AC12-49A9-B54B-21456E9139CB}
+		{1B65E7C5-1857-48DB-9F8D-07F67C39FA0B} = {537B0F0D-AC12-49A9-B54B-21456E9139CB}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {2F6B1794-D5DB-4788-A4E3-2786DA1F4359}
