+++ src/Hermodr.Publisher.DeliveryLog/LoggerExtensions.cs
@@ -18,4 +18,10 @@ static partial class LoggerExtensions
 
     [LoggerMessage(33004, LogLevel.Debug, "Delivery log error handler recorded failure for EventId={EventId}, Channel={Channel}")]
     public static partial void LogDeliveryLogErrorHandlerRecorded(this ILogger logger, string eventId, string? channel);
+
+    [LoggerMessage(33005, LogLevel.Warning, "Failed to delete old delivery log file: {FilePath}")]
+    public static partial void LogFailedToDeleteOldDeliveryLogFile(this ILogger logger, Exception ex, string filePath);
+
+    [LoggerMessage(33006, LogLevel.Warning, "Failed to clean up old delivery log files.")]
+    public static partial void LogFailedToCleanUpOldDeliveryLogFiles(this ILogger logger, Exception ex);
 }
